requirements.txt
# FABOT Podcast Studio - Dependencies
# Install with: pip install -r requirements.txt

# Core
fastapi>=0.135.0
uvicorn>=0.41.0
python-multipart>=0.0.22

# Database
sqlalchemy>=2.0.0

# Environment
python-dotenv>=1.2.0

# File Processing
pdfplumber>=0.11.0
PyPDF2>=3.0.0
python-docx>=1.2.0

# LLM & AI
aiohttp>=3.13.0
redis>=5.0.0

# Templates
jinja2>=3.1.0

# Validation
pydantic>=2.12.0

# Worker/Queue
arq>=0.27.0

# Audio Processing
soundfile>=0.13.0
pyloudnorm>=0.2.0
numpy>=1.26.0
scipy>=1.17.0

# Testing
pytest>=9.0.0
pytest-asyncio>=1.3.0
